# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateEntrepreneurRequest:
    properties:
      activity_status:
        type: boolean
      address:
        type: string
      charter_fund:
        type: integer
      dbibt_code:
        type: integer
      director_name:
        type: string
      email:
        type: string
      founders:
        type: string
      ifut_code:
        type: string
      inn:
        maxLength: 14
        minLength: 9
        type: string
      legal_form:
        type: string
      legal_name:
        type: string
      mhobt_code:
        type: string
      phone:
        maxLength: 32
        minLength: 9
        type: string
      registration_authority:
        type: string
      registration_date:
        type: string
      registration_number:
        type: string
    required:
    - email
    - inn
    - phone
    - registration_date
    type: object
  dto.EntrepreneurListResponse:
    properties:
      entrepreneurs:
        items:
          $ref: '#/definitions/dto.EntrepreneurResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  dto.EntrepreneurResponse:
    properties:
      activity_status:
        type: boolean
      address:
        type: string
      charter_fund:
        type: integer
      created_at:
        type: string
      dbibt_code:
        type: integer
      director_name:
        type: string
      email:
        type: string
      founders:
        type: string
      id:
        type: string
      ifut_code_id:
        type: string
      ifut_code_name:
        type: string
      inn_id:
        type: string
      inn_name:
        type: string
      legal_form:
        type: string
      legal_name:
        type: string
      mhobt_code:
        type: string
      phone:
        type: string
      registration_authority:
        type: string
      registration_date:
        type: string
      registration_number:
        type: string
      sqb_api_error:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.UpdateEntrepreneurRequest:
    properties:
      activity_status:
        type: boolean
      address:
        type: string
      charter_fund:
        type: integer
      dbibt_code:
        type: integer
      director_name:
        type: string
      email:
        type: string
      founders:
        type: string
      ifut_code:
        type: string
      legal_form:
        type: string
      legal_name:
        type: string
      mhobt_code:
        type: string
      phone:
        type: string
      registration_authority:
        type: string
      registration_date:
        type: string
      registration_number:
        type: string
    type: object
...
host: bank-app-back.compile-me.uz
info:
  contact: {}
  description: Authentication service for the Bank App
  title: Bank App API
  version: "1.0"
paths:
...
  /entrepreneurs:
    get:
      description: Returns a paginated list of entrepreneurs with optional filters
      parameters:
      - description: Search by legal name (partial match)
        in: query
        name: legal_name
        type: string
      - description: Search by INN name (partial match)
        in: query
        name: inn_name
        type: string
      - description: Filter by activity status
        in: query
        name: activity_status
        type: boolean
      - description: Search by director name (partial match)
        in: query
        name: director_name
        type: string
      - description: Filter from date (YYYY-MM-DD, e.g. 2024-01-01)
        in: query
        name: date_from
        type: string
      - description: Filter to date inclusive (YYYY-MM-DD, e.g. 2024-12-31)
        in: query
        name: date_to
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Entrepreneurs list
          schema:
            $ref: '#/definitions/dto.EntrepreneurListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all entrepreneurs
      tags:
      - entrepreneurs
    post:
      consumes:
      - application/json
      description: Creates a new entrepreneur. If the INN name exists, it binds the
        existing INN; otherwise creates a new one.
      parameters:
      - description: Entrepreneur details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEntrepreneurRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Entrepreneur created
          schema:
            $ref: '#/definitions/dto.EntrepreneurResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new entrepreneur
      tags:
      - entrepreneurs
  /entrepreneurs/{id}:
    delete:
      description: Deletes an entrepreneur by its UUID
      parameters:
      - description: Entrepreneur ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Entrepreneur deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Entrepreneur not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete entrepreneur
      tags:
      - entrepreneurs
    get:
      description: Returns an entrepreneur by its UUID
      parameters:
      - description: Entrepreneur ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Entrepreneur found
          schema:
            $ref: '#/definitions/dto.EntrepreneurResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Entrepreneur not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get entrepreneur by ID
      tags:
      - entrepreneurs
    put:
      consumes:
      - application/json
      description: Updates an entrepreneur's fields
      parameters:
      - description: Entrepreneur ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEntrepreneurRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Entrepreneur updated
          schema:
            $ref: '#/definitions/dto.EntrepreneurResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Entrepreneur not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update entrepreneur
      tags:
      - entrepreneurs
  /entrepreneurs/sqb-failed:
    get:
      description: Returns a paginated list of entrepreneurs that failed to send to
        SQB
      parameters:
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Failed entrepreneurs list
          schema:
            $ref: '#/definitions/dto.EntrepreneurListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get entrepreneurs with SQB errors
      tags:
      - entrepreneurs
  /entrepreneurs/sqb-retry:
    post:
      description: Fetches all entrepreneurs with sqb_api_error and resends them to
        SQB. Clears the error on success.
      produces:
      - application/json
      responses:
        "200":
          description: Retry results
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retry sending failed entrepreneurs to SQB
      tags:
      - entrepreneurs
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"